Deduplicate battle/menu objective condition implementations
_battle_condition.py and _menu_condition.py were identical except for the
word battle/menu in their space descriptions. The implementation now lives
once in _asm_condition.py; the two modules subclass it setting only
condition_type. The variants remain distinct classes deliberately:
_CachedFunction caches written routines per class, so merging them would
change the rom layout. Space descriptions are preserved exactly.
